refactor(embed_loop): extract policy scan and batch-write modules
Split reconcile/embed_loop.rs into orchestration (embed_loop),
embedding-policy scan/heal (policy_scan), and embed batch grouping
and remote-scoped writes (batch_write). Move reconstruct_file_text
unit tests alongside policy_scan; expose pub(crate) test helpers for
batch-write retry classification.
